Make the CI pass (#32)
* Exclude CLA signatures from pre-commit checks

The CLA JSON does not have an EOL at the end... ignore this.

* Fix the main Dockerfile

We were using poetry without first installing it – I've rectified this. The installation script is included in the repo to control what we are executing in CI – this is also recommended here: https://python-poetry.org/docs/#ci-recommendations

* docker-compose -> docker compose

This should fix Compose in GitHub Actions.