refactor: multiprocessing off in sc2_replaypack_processor
Currently the issue with multiprocessing is that the SC2InfoExtractorGo downloads the maps to get the localized map name into the final JSON file. Due to this there cannot be multiple processes running unless all of the required maps are download prior to executing the SC2InfoExtractorGo. This could be changed in the future if the first pass is ran with -download_maps_only